requirements.txt
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2018.8.24
cffi==1.12.3
Click==7.0
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
decorator==4.4.0
entrypoints==0.3
Flask==1.1.1
gevent==1.4.0
gunicorn==19.9.0
html5lib==1.0.1
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
MarkupSafe==1.1.1
msgpack==0.6.1
multipledispatch==0.6.0
nltk==3.4.4
nose==1.3.7
notebook==6.0.0
numba==0.45.0
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
openpyxl==2.6.2
packaging==19.0
path.py==12.0.1
pathlib2==2.3.4
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.2
QtPy==1.8.0
requests==2.22.0
requests-oauthlib==1.2.0
seaborn==0.9.0
soupsieve==1.8
SQLAlchemy==1.3.5
statsmodels==0.10.0
sympy==1.4
tables==3.5.2
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.8
zict==1.0.0
zipp==0.5.1
